requirements.txt
asgiref==3.7.2
Babel==2.12.1
certifi==2023.7.22
charset-normalizer==3.2.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.22
django-cors-headers==4.2.0
django-environ==0.11.2
django-js-asset==2.1.0
django-money==3.2.0
django-mptt==0.14.0
django-mptt-admin==2.5.1
django-polymorphic==3.1.0
django-rest-swagger==2.2.0
django-sharding==5.2.0
django-tenants==3.5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.7
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
openapi-codec==1.3.2
packaging==23.1
Pillow==10.0.0
psycopg2-binary==2.9.6
py-moneyed==3.0
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
simplejson==3.19.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==2.0.4
whitenoise==6.5.0
